+++ flag/service/cnr/cnr.go
@@ -1,7 +1,0 @@
-package cnr
-
-// CNR is a data structure to hold CNR specific command line configuration flags.
-type CNR struct {
-	Address      string
-	Organization string
-}