+++ integration/key/release_name.go
@@ -4,10 +4,6 @@ func ChartOperatorReleaseName() string {
 	return "chart-operator"
 }
 
-func CustomResourceReleaseName() string {
-	return "apiextensions-chart-e2e-chart"
-}
-
 func TestAppReleaseName() string {
 	return "test-app"
 }