+++ integration/setup/config.go
@@ -1,10 +1,8 @@
 package setup
 
 import (
-	"github.com/giantswarm/e2e-harness/pkg/framework/resource"
 	"github.com/giantswarm/e2e-harness/pkg/release"
 	"github.com/giantswarm/e2esetup/chart/env"
-	"github.com/giantswarm/e2esetup/k8s"
 	"github.com/giantswarm/helmclient"
 	"github.com/giantswarm/k8sclient"
 	"github.com/giantswarm/microerror"
@@ -18,11 +16,10 @@ const (
 
 type Config struct {
 	HelmClient *helmclient.Client
-	K8s        *k8s.Setup
+	K8s        *k8sclient.Setup
 	K8sClients *k8sclient.Clients
 	Logger     micrologger.Logger
 	Release    *release.Release
-	Resource   *resource.Resource
 }
 
 func NewConfig() (Config, error) {
@@ -52,14 +49,14 @@ func NewConfig() (Config, error) {
 		}
 	}
 
-	var k8sSetup *k8s.Setup
+	var k8sSetup *k8sclient.Setup
 	{
-		c := k8s.SetupConfig{
+		c := k8sclient.SetupConfig{
 			Clients: cpK8sClients,
 			Logger:  logger,
 		}
 
-		k8sSetup, err = k8s.NewSetup(c)
+		k8sSetup, err = k8sclient.NewSetup(c)
 		if err != nil {
 			return Config{}, microerror.Mask(err)
 		}
@@ -68,10 +65,11 @@ func NewConfig() (Config, error) {
 	var helmClient *helmclient.Client
 	{
 		c := helmclient.Config{
-			Logger:          logger,
-			K8sClient:       cpK8sClients.K8sClient(),
-			RestConfig:      cpK8sClients.RestConfig(),
-			TillerNamespace: tillerNamespace,
+			Logger:               logger,
+			K8sClient:            cpK8sClients.K8sClient(),
+			RestConfig:           cpK8sClients.RESTConfig(),
+			TillerNamespace:      tillerNamespace,
+			TillerUpgradeEnabled: true,
 		}
 		helmClient, err = helmclient.New(c)
 		if err != nil {
@@ -97,27 +95,12 @@ func NewConfig() (Config, error) {
 		}
 	}
 
-	var newResource *resource.Resource
-	{
-		c := resource.Config{
-			Logger:     logger,
-			HelmClient: helmClient,
-			Namespace:  namespace,
-		}
-		newResource, err = resource.New(c)
-		if err != nil {
-			return Config{}, microerror.Mask(err)
-		}
-	}
-
 	c := Config{
 		HelmClient: helmClient,
 		K8s:        k8sSetup,
 		K8sClients: cpK8sClients,
 		Logger:     logger,
 		Release:    newRelease,
-		// Resource is deprecated and used by legacy chartconfig tests.
-		Resource: newResource,
 	}
 
 	return c, nil