+++ service/collector/collector.go
@@ -7,7 +7,6 @@ const (
 
 const (
 	labelChart         = "chart_name"
-	labelChannel       = "channel_name"
 	labelRelease       = "release_name"
 	labelReleaseStatus = "release_status"
 	labelNamespace     = "namespace"