+++ service/collector/set.go
@@ -1,17 +1,15 @@
 package collector
 
 import (
-	"github.com/giantswarm/apiextensions/pkg/clientset/versioned"
 	"github.com/giantswarm/exporterkit/collector"
 	"github.com/giantswarm/helmclient"
+	"github.com/giantswarm/k8sclient"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
-	"k8s.io/client-go/kubernetes"
 )
 
 type SetConfig struct {
-	G8sClient  versioned.Interface
-	K8sClient  kubernetes.Interface
+	K8sClient  k8sclient.Interface
 	HelmClient *helmclient.Client
 	Logger     micrologger.Logger
 
@@ -26,9 +24,6 @@ type Set struct {
 }
 
 func NewSet(config SetConfig) (*Set, error) {
-	if config.G8sClient == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.G8sClient must not be empty", config)
-	}
 	if config.K8sClient == nil {
 		return nil, microerror.Maskf(invalidConfigError, "%T.K8sClient must not be empty", config)
 	}
@@ -45,14 +40,29 @@ func NewSet(config SetConfig) (*Set, error) {
 
 	var err error
 
-	var chartConfigResourceCollector *ChartConfigResource
+	var orphanConfigMapCollector *OrphanConfigMap
 	{
-		c := ChartConfigResourceConfig{
-			G8sClient: config.G8sClient,
+		c := OrphanConfigMapConfig{
+			G8sClient: config.K8sClient.G8sClient(),
+			K8sClient: config.K8sClient.K8sClient(),
 			Logger:    config.Logger,
 		}
 
-		chartConfigResourceCollector, err = NewChartConfigResource(c)
+		orphanConfigMapCollector, err = NewOrphanConfigMap(c)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+	}
+
+	var orphanSecretCollector *OrphanSecret
+	{
+		c := OrphanSecretConfig{
+			G8sClient: config.K8sClient.G8sClient(),
+			K8sClient: config.K8sClient.K8sClient(),
+			Logger:    config.Logger,
+		}
+
+		orphanSecretCollector, err = NewOrphanSecret(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -61,8 +71,8 @@ func NewSet(config SetConfig) (*Set, error) {
 	var tillerMaxHistoryCollector *TillerMaxHistory
 	{
 		c := TillerMaxHistoryConfig{
-			G8sClient: config.G8sClient,
-			K8sClient: config.K8sClient,
+			G8sClient: config.K8sClient.G8sClient(),
+			K8sClient: config.K8sClient.K8sClient(),
 			Logger:    config.Logger,
 
 			TillerNamespace: config.TillerNamespace,
@@ -77,7 +87,7 @@ func NewSet(config SetConfig) (*Set, error) {
 	var tillerReachableCollector *TillerReachable
 	{
 		c := TillerReachableConfig{
-			G8sClient:  config.G8sClient,
+			G8sClient:  config.K8sClient.G8sClient(),
 			HelmClient: config.HelmClient,
 			Logger:     config.Logger,
 
@@ -93,7 +103,7 @@ func NewSet(config SetConfig) (*Set, error) {
 	var tillerRunningPodsCollector *TillerRunningPods
 	{
 		c := TillerRunningPodsConfig{
-			G8sClient:  config.G8sClient,
+			G8sClient:  config.K8sClient.G8sClient(),
 			HelmClient: config.HelmClient,
 			Logger:     config.Logger,
 
@@ -110,7 +120,8 @@ func NewSet(config SetConfig) (*Set, error) {
 	{
 		c := collector.SetConfig{
 			Collectors: []collector.Interface{
-				chartConfigResourceCollector,
+				orphanConfigMapCollector,
+				orphanSecretCollector,
 				tillerMaxHistoryCollector,
 				tillerReachableCollector,
 				tillerRunningPodsCollector,