+++ service/controller/chart/v1/resource/release/create.go
@@ -3,12 +3,14 @@ package release
 import (
 	"context"
 	"fmt"
+	"time"
 
 	"github.com/giantswarm/helmclient"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/operatorkit/controller/context/resourcecanceledcontext"
 	"k8s.io/helm/pkg/helm"
 
+	"github.com/giantswarm/chart-operator/service/controller/chart/v1/controllercontext"
 	"github.com/giantswarm/chart-operator/service/controller/chart/v1/key"
 )
 
@@ -17,6 +19,11 @@ func (r *Resource) ApplyCreateChange(ctx context.Context, obj, createChange inte
 	if err != nil {
 		return microerror.Mask(err)
 	}
+	cc, err := controllercontext.FromContext(ctx)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
 	releaseState, err := toReleaseState(createChange)
 	if err != nil {
 		return microerror.Mask(err)
@@ -30,11 +37,28 @@ func (r *Resource) ApplyCreateChange(ctx context.Context, obj, createChange inte
 
 		tarballPath, err := r.helmClient.PullChartTarball(ctx, tarballURL)
 		if helmclient.IsPullChartFailedError(err) {
-			r.logger.LogCtx(ctx, "level", "warning", "message", "pulling chart failed", "stack", microerror.Stack(err))
+			reason := fmt.Sprintf("pulling chart %#q failed", tarballURL)
+			addStatusToContext(cc, reason, releaseNotInstalledStatus)
 
-			resourcecanceledcontext.SetCanceled(ctx)
+			r.logger.LogCtx(ctx, "level", "warning", "message", reason, "stack", microerror.JSON(err))
 			r.logger.LogCtx(ctx, "level", "debug", "message", "canceling resource")
+			resourcecanceledcontext.SetCanceled(ctx)
+			return nil
+		} else if helmclient.IsPullChartNotFound(err) {
+			reason := fmt.Sprintf("chart %#q not found", tarballURL)
+			addStatusToContext(cc, reason, releaseNotInstalledStatus)
 
+			r.logger.LogCtx(ctx, "level", "warning", "message", reason, "stack", microerror.JSON(err))
+			r.logger.LogCtx(ctx, "level", "debug", "message", "canceling resource")
+			resourcecanceledcontext.SetCanceled(ctx)
+			return nil
+		} else if helmclient.IsPullChartTimeout(err) {
+			reason := fmt.Sprintf("timeout pulling %#q", tarballURL)
+			addStatusToContext(cc, reason, releaseNotInstalledStatus)
+
+			r.logger.LogCtx(ctx, "level", "warning", "message", reason, "stack", microerror.JSON(err))
+			r.logger.LogCtx(ctx, "level", "debug", "message", "canceling resource")
+			resourcecanceledcontext.SetCanceled(ctx)
 			return nil
 		} else if err != nil {
 			return microerror.Mask(err)
@@ -47,21 +71,49 @@ func (r *Resource) ApplyCreateChange(ctx context.Context, obj, createChange inte
 			}
 		}()
 
-		// We need to pass the ValueOverrides option to make the install process
-		// use the default values and prevent errors on nested values.
-		err = r.helmClient.InstallReleaseFromTarball(ctx, tarballPath, ns, helm.ReleaseName(releaseState.Name), helm.ValueOverrides(releaseState.ValuesYAML))
+		ch := make(chan error)
+
+		// We create the helm release but with a short timeout so we don't
+		// block reconciling other CRs. This gives time to make the port
+		// forwarding connection to the Tiller API.
+		//
+		// If we do timeout the install will continue in the background.
+		// We will check the progress in the next reconciliation loop.
+		go func() {
+			// We need to pass the ValueOverrides option to make the install process
+			// use the default values and prevent errors on nested values.
+			err = r.helmClient.InstallReleaseFromTarball(ctx, tarballPath, ns, helm.ReleaseName(releaseState.Name), helm.ValueOverrides(releaseState.ValuesYAML))
+			close(ch)
+		}()
+
+		select {
+		case <-ch:
+			// Fall through.
+		case <-time.After(3 * time.Second):
+			r.logger.LogCtx(ctx, "level", "debug", "message", "release still being created")
+			r.logger.LogCtx(ctx, "level", "debug", "message", "canceling resource")
+			return nil
+		}
+
 		if err != nil {
+			reason := err.Error()
+			r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("helm release %#q failed", releaseState.Name), "stack", microerror.JSON(err))
+
 			releaseContent, err := r.helmClient.GetReleaseContent(ctx, releaseState.Name)
 			if helmclient.IsReleaseNotFound(err) {
-				r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("helm release %#q not found", releaseContent.Name))
+				reason = fmt.Sprintf("helm error: (%s)", reason)
+				addStatusToContext(cc, reason, releaseNotInstalledStatus)
+
 				r.logger.LogCtx(ctx, "level", "debug", "message", "canceling resource")
 				resourcecanceledcontext.SetCanceled(ctx)
 				return nil
 			} else if err != nil {
 				return microerror.Mask(err)
 			}
+
+			// Release is failed so the status resource will check the Helm release.
 			if releaseContent.Status == helmFailedStatus {
-				r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("failed to update release %#q", releaseContent.Name))
+				r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("failed to create release %#q", releaseContent.Name))
 				r.logger.LogCtx(ctx, "level", "debug", "message", "canceling resource")
 				resourcecanceledcontext.SetCanceled(ctx)
 				return nil